added another folder for splitting the components inside Navbar.js
