+++ client/src/components/FoodPostCard.css
@@ -9,12 +9,9 @@
 } */
 
 .post {
-  width: 100%;
-  border-radius: 10px;
-  box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
-  margin-bottom: 50px;
-  background-color: #cdb2ab;
-  margin-top: 10px;
+  align-items: center;
+  display: flex;
+  justify-content: center;
 }
 
 .post-wrapper {
@@ -22,6 +19,12 @@
   overflow: hidden;
   display: fixed;
   margin-left: 0;
+  width: 80%;
+  border-radius: 10px;
+  box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+  margin-bottom: 50px;
+  background-color: #cdb2ab;
+  margin-top: 10px;
 }
 
 .post-top {
@@ -49,18 +52,14 @@
   margin: 0 10px;
 }
 
+.post-location {
+  font-size: 15px;
+  font-weight: 500;
+  margin-right: 5px;
+}
+
 .post-center {
   margin: 20px 0;
-}
-
-.post-description {
-  align-items: center;
-  display: flex;
-  justify-content: center;
-}
-
-.post-description-title {
-  margin-right: 20px;
 }
 
 .image-box {
@@ -71,13 +70,16 @@
 
 .post-image {
   margin-top: 20px;
-  width: 700px;
+  width: 100%;
   border-radius: 10px;
   box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+
+  height: auto;
 }
 
 .going-button {
   display: flex;
   align-items: center;
   justify-content: flex-start;
+  border-radius: 10px;
 }
