+++ client/src/components/FoodPostCard.js
@@ -29,7 +29,7 @@ function FoodPostCard() {
             I would like to give away a vegan carrot cake that I baked this
             morning. Enough for four people
           </p>
-          <div className = "image-box">
+          <div className="image-box">
             <img
               className="post-image"
               src="https://img.delicious.com.au/Kl8Gm8gM/del/2017/05/spiced-carrot-cake-with-labneh-icing-46402-2.jpg"
