+++ client/src/components/LoadingButton.css
@@ -10,7 +10,7 @@
 
 .load-button {
   padding: 20px;
-  width: 100%;
+  width: 85%;
   background-color: #a37774 !important;
   border-radius: 10px;
 }
