+++ client/src/components/Login.css
@@ -1,9 +1,10 @@
 * {
   box-sizing: border-box;
+  
 }
 
 #container {
-  background: #f6f5f7;
+  
   display: flex;
   justify-content: center;
   align-items: center;
@@ -11,7 +12,13 @@
   font-family: monospace, sans-serif;
   height: 100vh;
   margin: -20px 0 50px;
-  background-color: #f7ede2;
+  background-image: url("https://cdn.dribbble.com/users/941157/screenshots/15513050/media/6413f9fe3d03bd17202fcaebd2c8c2f9.jpeg?compress=1&resize=1200x900");
+}
+
+#container .form-container {
+  border-radius: 10px;
+  box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+  background-color: #f7ede2 !important;
 }
 
 h1 {
