--- client/src/components/Navbar.css
@@ -1,39 +1,0 @@
-.navbar {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  position: sticky;
-  height: 40px;
-  top: 0;
-  z-index: 1000;
-  width: 100%;
-  background-color: #84a59d;
-}
-
-.homepage-logo {
-  height: 45px;
-  object-fit: contain;
-  margin-top: 4.5px;
-  margin-left: 4.5px;
-  border-radius: 50%;
-}
-
-.homepage-input-search {
-  display: flex;
-  align-items: center;
-  width: 40%;
-  float: left;
-}
-
-.navbar > .homepage-input-search > input {
-  width: 400px;
-  height: 25px;
-  margin-left: -3%;
-  display: flex;
-  align-items: inherit;
-}
-
-.homepage-input-search > input {
-  flex: 1;
-  border: none;
-}
