+++ client/src/components/ReservedPopup.css
@@ -1,15 +1,13 @@
 .popup {
   border: 2px solid #ccc;
   padding: 10px;
-  width: 100%;
+  width: 400%;
   border-radius: 10px;
-  box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
-  margin: 30px 30px;
+  box-shadow: 0px 0px 16px -8px #000000ad;
+  margin: 30px 5px 30px 30px;
 }
 
 .popup-inner {
- 
-  object-fit: contain;
   background-color: rgb(246, 225, 225);
   padding: 5px;
   margin-bottom: 1em;
