+++ client/src/pages/ExtraFoodCardOne.js
@@ -1,10 +1,8 @@
 import React from "react";
 import { useState } from "react";
-//import "./FoodPostCard.css";
-import "../components/FoodPostCard.css";
 import ReservedPopup from "../components/ReservedPopup";
-//import "./ReservedPopup.css";
 import "../components/ReservedPopup.css";
+import "../components/FoodPostCard.css";
 
 function ExtraFoodCardOne() {
   const [buttonPopup, setButtonPopup] = useState(false);
@@ -35,7 +33,7 @@ function ExtraFoodCardOne() {
             alt="pizza"
           ></img>
         </div>
-        <div className="going-bottom">
+        <div className="going-button">
           <button
             className="going-post-food-btn"
             onClick={() => setButtonPopup(true)}
